Refactor of area based analysis data creation. Remove functions from plymouth script and put into a utils
